draft 0 of the tablet code (proof of concept UI repaint + socket in separate threads)
works in both Java and Android mode. Demonstrates simple thread that
fetches content of a URL while the UI thread keeps drawing flawlessly.
(just make sure that the url in the code points to something actually
being served). Touch/click on the rectangle: the content of the web
page in the url should appear while the ui is smoothly updated.